+++ libs/docs/core/date-picker/examples/date-picker-format-example.component.ts
@@ -39,10 +39,42 @@ export const CUSTOM_FD_DATETIME_FORMATS: DateTimeFormats = {
         <br />
         <div>Selected Date: {{ date | dateFormat }}</div>
         <br />
+        <fd-date-picker
+            [(ngModel)]="date"
+            [customDateTimeFormat]="componentSpecificDateTimeFormat"
+            placeholder="dd-mm-yyyy"
+        ></fd-date-picker>
+        <br />
+        <div>Selected Date: {{ date | dateFormat }}</div>
+        <br />
+        <fd-date-picker [(ngModel)]="dates" placeholder="dd-mm-yyyy"></fd-date-picker>
+        <br />
+        <div>
+            Selected Dates:<br />
+            @for (date of dates; track date) {
+                {{ date.toDateString() || 'null' }}<br />
+            }
+        </div>
+        <br />
         <fd-date-picker placeholder="mm/dd/yy to mm/dd/yy" type="range" [(ngModel)]="selectedRange"></fd-date-picker>
         <br />
         <div>Selected First Date: {{ selectedRange?.start | dateFormat }}</div>
         <div>Selected Last Date: {{ selectedRange?.end | dateFormat }}</div>
+
+        <br />
+        <fd-date-picker
+            placeholder="mm/dd/yy to mm/dd/yy"
+            type="range"
+            [(ngModel)]="selectedRanges"
+            [allowMultipleSelection]="true"
+        ></fd-date-picker>
+        <div>
+            Selected Date Ranges: <br />
+            @for (dateRange of selectedRanges; track dateRange) {
+                {{ (dateRange?.start?.toDateString() || 'null') + ' - ' + (dateRange?.end?.toDateString() || 'null') }}
+                <br />
+            }
+        </div>
     `,
     providers: [
         {
@@ -59,11 +91,31 @@ export const CUSTOM_FD_DATETIME_FORMATS: DateTimeFormats = {
 })
 export class DatePickerFormatExampleComponent {
     date: FdDate;
+    dates: Nullable<FdDate[]> = [
+        new FdDate(2019, 9, 1),
+        new FdDate(2019, 9, 2),
+        new FdDate(2019, 9, 3),
+        new FdDate(2019, 9, 4)
+    ];
     selectedRange: Nullable<DateRange<FdDate>>;
+    selectedRanges: Nullable<Array<DateRange<FdDate>>>;
+    componentSpecificDateTimeFormat = {
+        year: '2-digit',
+        month: 'long',
+        day: '2-digit'
+    };
 
     constructor(private datetimeAdapter: DatetimeAdapter<FdDate>) {
         const today = this.datetimeAdapter.today();
+        const other1 = new FdDate(2024, 8, 1);
+        const other2 = new FdDate(2024, 8, 10);
+        const other3 = new FdDate(2024, 8, 20);
         this.date = today;
         this.selectedRange = new DateRange(today, this.datetimeAdapter.addCalendarDays(today, 1));
+        this.selectedRanges = [
+            new DateRange(other1, this.datetimeAdapter.addCalendarDays(other1, 5)),
+            new DateRange(other2, this.datetimeAdapter.addCalendarDays(other2, 5)),
+            new DateRange(other3, this.datetimeAdapter.addCalendarDays(other3, 5))
+        ];
     }
 }